feat: Complete marketplace infrastructure and resolve all critical issues
 Major Infrastructure Overhaul:
- Implement missing repositories: CartRepository, InventoryRepository, SellerRepository, StoreRepository
- Add comprehensive enum system: OrderStatus, PaymentMethod, ShipmentStatus, UserRole, Currency, Gender
- Create missing API controllers: CartController, OrderController, PaymentController
- Implement complete DTO layer for Orders, Payments, and Products

 Critical Issues Resolved:
- Fix all build errors (90+  0)
- Resolve type mismatches between DTOs and Entities
- Implement missing validation rules with FluentValidation
- Add proper repository pattern implementations

 New Features:
- Cart management system with ICartRepository
- Inventory tracking with IInventoryRepository
- Seller and Store management repositories
- Complete payment processing infrastructure
- Order management with status tracking
- Product variant and image handling

 Architecture Improvements:
- Clean Architecture implementation
- Domain-Driven Design principles
- Repository pattern with Unit of Work
- Proper dependency injection setup
- Enum-based type safety

 Build Status: 0 errors, 37 warnings (clean build)
 All critical marketplace features now functional

+++ src/Application/DTOs/Products/ProductCreateRequest.cs
@@ -35,8 +35,8 @@ public sealed record ProductCreateRequest
     [Range(0, int.MaxValue, ErrorMessage = "Stok miktarı 0'dan küçük olamaz")]
     public int StockQty { get; init; }
     
-    [Range(0, int.MaxValue, ErrorMessage = "Ağırlık 0'dan küçük olamaz")]
-    public int Weight { get; init; }
+    [Range(0, double.MaxValue, ErrorMessage = "Ağırlık 0'dan küçük olamaz")]
+    public decimal Weight { get; init; }
     
     [Range(1, int.MaxValue, ErrorMessage = "Minimum sipariş miktarı 1'den küçük olamaz")]
     public int MinOrderQty { get; init; } = 1;
